pyproject.toml
[tool.poetry]
name = "duality"
version = "0.1.0"
description = ""
authors = ["Matt Kramer <mkramer@anaconda.com>"]

[tool.poetry.dependencies]
python = "^3.9"

azure-digitaltwins-core = "^1.1.0"
azure-identity = "^1.7.1"
python-dotenv = "^0.19.2"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
black = "21.11b1"
mypy = "^0.910"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
